+++ src/php/Lsp/Application/Handler/DefinitionHandler.php
@@ -8,9 +8,9 @@ use Phel\Api\ApiFacade;
 use Phel\Api\Transfer\Definition;
 use Phel\Api\Transfer\ProjectIndex;
 use Phel\Lsp\Application\Convert\LocationConverter;
+use Phel\Lsp\Application\HandlerInterface;
 use Phel\Lsp\Application\Rpc\ParamsExtractor;
 use Phel\Lsp\Application\Session\Session;
-use Phel\Lsp\Domain\HandlerInterface;
 
 use function str_contains;