+++ src/php/Lsp/Application/Handler/DidOpenHandler.php
@@ -5,9 +5,9 @@ declare(strict_types=1);
 namespace Phel\Lsp\Application\Handler;
 
 use Phel\Lsp\Application\Diagnostics\DiagnosticPublisher;
+use Phel\Lsp\Application\HandlerInterface;
 use Phel\Lsp\Application\Rpc\ParamsExtractor;
 use Phel\Lsp\Application\Session\Session;
-use Phel\Lsp\Domain\HandlerInterface;
 
 final readonly class DidOpenHandler implements HandlerInterface
 {