+++ src/php/Lsp/Application/Handler/WorkspaceSymbolHandler.php
@@ -6,8 +6,8 @@ namespace Phel\Lsp\Application\Handler;
 
 use Phel\Api\Transfer\ProjectIndex;
 use Phel\Lsp\Application\Convert\SymbolInformationBuilder;
+use Phel\Lsp\Application\HandlerInterface;
 use Phel\Lsp\Application\Session\Session;
-use Phel\Lsp\Domain\HandlerInterface;
 
 use function is_string;
 use function str_contains;